// src/generated/stream_list.ts
export enum LiveChatMessageType {
  INVALID_TYPE = 0,
  TEXT_MESSAGE_EVENT = 1,
  TOMBSTONE = 2,
  CHAT_ENDED_EVENT = 4,
  SPONSOR_ONLY_MODE_STARTED_EVENT = 5,
  SPONSOR_ONLY_MODE_ENDED_EVENT = 6,
  NEW_SPONSOR_EVENT = 7,
  USER_BANNED_EVENT = 10,
  SUPER_CHAT_EVENT = 15,
  SUPER_STICKER_EVENT = 16,
  MEMBER_MILESTONE_CHAT_EVENT = 17,
  MEMBERSHIP_GIFTING_EVENT = 18,
  GIFT_MEMBERSHIP_RECEIVED_EVENT = 19,
  POLL_EVENT = 20,
  GIFT_EVENT = 21,
  UNRECOGNIZED = -1,
}
// ...
export interface LiveChatMessage {
  id?: string | undefined;
  snippet?: LiveChatMessageSnippet | undefined;
  authorDetails?: LiveChatMessageAuthorDetails | undefined;
}

export interface LiveChatMessageAuthorDetails {
  channelId?: string | undefined;
  displayName?: string | undefined;
  isVerified?: boolean | undefined;
  isChatOwner?: boolean | undefined;
  isChatSponsor?: boolean | undefined;
  isChatModerator?: boolean | undefined;
}

export interface LiveChatMessageSnippet {
  type?: LiveChatMessageType | undefined;
  publishedAt?: string | undefined;
  displayMessage?: string | undefined;
  hasDisplayContent?: boolean | undefined;
}
// ...
export const LiveChatMessage: MessageFns<LiveChatMessage> = {
  encode(message: LiveChatMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== undefined && message.id !== "") {
      writer.uint32(810).string(message.id);
    }
    if (message.snippet !== undefined) {
      LiveChatMessageSnippet.encode(message.snippet, writer.uint32(18).fork()).join();
    }
    if (message.authorDetails !== undefined) {
      LiveChatMessageAuthorDetails.encode(message.authorDetails, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LiveChatMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLiveChatMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 101: {
          if (tag !== 810) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.snippet = LiveChatMessageSnippet.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.authorDetails = LiveChatMessageAuthorDetails.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LiveChatMessage {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      snippet: isSet(object.snippet) ? LiveChatMessageSnippet.fromJSON(object.snippet) : undefined,
      authorDetails: isSet(object.authorDetails)
        ? LiveChatMessageAuthorDetails.fromJSON(object.authorDetails)
        : isSet(object.author_details)
        ? LiveChatMessageAuthorDetails.fromJSON(object.author_details)
        : undefined,
    };
  },

  toJSON(message: LiveChatMessage): unknown {
    const obj: any = {};
    if (message.id !== undefined && message.id !== "") {
      obj.id = message.id;
    }
    if (message.snippet !== undefined) {
      obj.snippet = LiveChatMessageSnippet.toJSON(message.snippet);
    }
    if (message.authorDetails !== undefined) {
      obj.authorDetails = LiveChatMessageAuthorDetails.toJSON(message.authorDetails);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LiveChatMessage>, I>>(base?: I): LiveChatMessage {
    return LiveChatMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LiveChatMessage>, I>>(object: I): LiveChatMessage {
    const message = createBaseLiveChatMessage();
    message.id = object.id ?? "";
    message.snippet = (object.snippet !== undefined && object.snippet !== null)
      ? LiveChatMessageSnippet.fromPartial(object.snippet)
      : undefined;
    message.authorDetails = (object.authorDetails !== undefined && object.authorDetails !== null)
      ? LiveChatMessageAuthorDetails.fromPartial(object.authorDetails)
      : undefined;
    return message;
  },
};
// ...
export const LiveChatMessageSnippet: MessageFns<LiveChatMessageSnippet> = {
  encode(message: LiveChatMessageSnippet, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== undefined && message.type !== 0) {
      writer.uint32(8).int32(message.type);
    }
    if (message.publishedAt !== undefined && message.publishedAt !== "") {
      writer.uint32(34).string(message.publishedAt);
    }
    if (message.displayMessage !== undefined && message.displayMessage !== "") {
      writer.uint32(130).string(message.displayMessage);
    }
    if (message.hasDisplayContent !== undefined && message.hasDisplayContent !== false) {
      writer.uint32(136).bool(message.hasDisplayContent);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LiveChatMessageSnippet {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLiveChatMessageSnippet();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.publishedAt = reader.string();
          continue;
        }
        case 16: {
          if (tag !== 130) {
            break;
          }

          message.displayMessage = reader.string();
          continue;
        }
        case 17: {
          if (tag !== 136) {
            break;
          }

          message.hasDisplayContent = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LiveChatMessageSnippet {
    return {
      type: isSet(object.type) ? liveChatMessageTypeFromJSON(object.type) : 0,
      publishedAt: isSet(object.publishedAt)
        ? globalThis.String(object.publishedAt)
        : isSet(object.published_at)
        ? globalThis.String(object.published_at)
        : "",
      displayMessage: isSet(object.displayMessage)
        ? globalThis.String(object.displayMessage)
        : isSet(object.display_message)
        ? globalThis.String(object.display_message)
        : "",
      hasDisplayContent: isSet(object.hasDisplayContent)
        ? globalThis.Boolean(object.hasDisplayContent)
        : isSet(object.has_display_content)
        ? globalThis.Boolean(object.has_display_content)
        : false,
    };
  },

  toJSON(message: LiveChatMessageSnippet): unknown {
    const obj: any = {};
    if (message.type !== undefined && message.type !== 0) {
      obj.type = liveChatMessageTypeToJSON(message.type);
    }
    if (message.publishedAt !== undefined && message.publishedAt !== "") {
      obj.publishedAt = message.publishedAt;
    }
    if (message.displayMessage !== undefined && message.displayMessage !== "") {
      obj.displayMessage = message.displayMessage;
    }
    if (message.hasDisplayContent !== undefined && message.hasDisplayContent !== false) {
      obj.hasDisplayContent = message.hasDisplayContent;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LiveChatMessageSnippet>, I>>(base?: I): LiveChatMessageSnippet {
    return LiveChatMessageSnippet.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LiveChatMessageSnippet>, I>>(object: I): LiveChatMessageSnippet {
    const message = createBaseLiveChatMessageSnippet();
    message.type = object.type ?? 0;
    message.publishedAt = object.publishedAt ?? "";
    message.displayMessage = object.displayMessage ?? "";
    message.hasDisplayContent = object.hasDisplayContent ?? false;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}